refactor: extract _DelimitedConverter base from DelimitedCollection
Move the collection_delimiter class variable, the field-classification scan,
and the shared field validator/serializer into a non-exported
_DelimitedConverter base; DelimitedCollection becomes a thin subclass that
carries only the user-facing docstring.

This sets up the single field_serializer("*") on a shared base, which a second
delimited converter can build on without colliding (Pydantic permits only one
serializer per field across the MRO). The base dispatches on the classification
sets directly, so the test-only _is_collection_field predicate is dropped.

Pure refactor; no behavior change.

Relates to #76.
